+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:hexagon/grid/hexagon_offset_grid.dart';
 import 'package:hexagon/hexagon_type.dart';
 import 'package:hexagon/hexagon_widget.dart';
 
@@ -29,197 +30,150 @@ class MyHomePage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     var size = MediaQuery.of(context).size;
+
     return DefaultTabController(
       length: 3,
+      initialIndex: 0,
       child: Scaffold(
         appBar: AppBar(
           bottom: TabBar(
             tabs: [
-              Tab(
-                text: 'Flat',
-              ),
-              Tab(
-                text: 'Pointy',
-              ),
-              Tab(text: 'More'),
+              Tab(text: 'V-Grid'),
+              Tab(text: 'H-Grid'),
+              Tab(text: 'Other'),
             ],
           ),
           title: Text(title),
         ),
         body: TabBarView(
           children: [
-            _buildFlatExample(size),
-            _buildPointyExample(size),
-            _buildMore(),
+            _buildVerticalGrid(size),
+            _buildHorizontalGrid(size),
+            _buildMore(size),
           ],
         ),
       ),
     );
   }
 
-  Widget _buildFlatExample(Size size) {
-    var w = size.width > 200.0 ? 200.0 : size.width / 2;
-    var h = 200.0;
+  Widget _buildHorizontalGrid(Size size) {
+    return SingleChildScrollView(
+      scrollDirection: Axis.horizontal,
+      child: HexagonOffsetGrid.oddPointy(
+        color: Colors.black54,
+        columns: 9,
+        rows: 4,
+        buildHexagon: (col, row) => row.isOdd && col.isOdd
+            ? null
+            : HexagonWidget.template(
+                elevation: col.toDouble(),
+                color: col == 1 || row == 1 ? Colors.lightBlue.shade200 : null,
+                child: Text('$col, $row'),
+              ),
+      ),
+    );
+  }
+
+  Widget _buildVerticalGrid(Size size) {
     return SingleChildScrollView(
       child: Column(
+        crossAxisAlignment: CrossAxisAlignment.stretch,
         children: [
-          Padding(
-            padding: EdgeInsets.all(8.0),
-            child: HexagonWidget.flat(
-              width: w,
-              child: AspectRatio(
-                aspectRatio: HexagonType.FLAT.ratio,
-                child: Image.asset(
-                  'assets/bee.jpg',
-                  fit: BoxFit.fitHeight,
+          HexagonOffsetGrid.evenFlat(
+            color: Colors.yellow.shade100,
+            columns: 5,
+            rows: 10,
+            buildHexagon: (col, row) => HexagonWidget.template(
+              color: row.isEven ? Colors.yellow : Colors.orangeAccent,
+              elevation: 2,
+            ),
+            buildChild: (col, row) => Text('$col, $row'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildMore(Size size) {
+    var padding = 8.0;
+    var w = (size.width - 4 * padding) / 2;
+    var h = 150.0;
+    return SingleChildScrollView(
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Padding(
+                padding: EdgeInsets.all(padding),
+                child: HexagonWidget.flat(
+                  width: w,
+                  child: AspectRatio(
+                    aspectRatio: HexagonType.FLAT.ratio,
+                    child: Image.asset(
+                      'assets/bee.jpg',
+                      fit: BoxFit.fitHeight,
+                    ),
+                  ),
                 ),
               ),
-            ),
+              Padding(
+                padding: EdgeInsets.all(padding),
+                child: HexagonWidget.pointy(
+                  width: w,
+                  child: AspectRatio(
+                    aspectRatio: HexagonType.POINTY.ratio,
+                    child: Image.asset(
+                      'assets/tram.jpg',
+                      fit: BoxFit.fitWidth,
+                    ),
+                  ),
+                ),
+              ),
+            ],
+          ),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              Padding(
+                padding: EdgeInsets.all(8.0),
+                child: HexagonWidget.flat(
+                  height: h,
+                  color: Colors.orangeAccent,
+                  child: Text('flat\nheight: ${h.toStringAsFixed(2)}'),
+                ),
+              ),
+              Padding(
+                padding: EdgeInsets.all(8.0),
+                child: HexagonWidget.pointy(
+                  height: h,
+                  color: Colors.red,
+                  child: Text('pointy\nheight: ${h.toStringAsFixed(2)}'),
+                ),
+              ),
+            ],
           ),
           Padding(
-            padding: EdgeInsets.all(8.0),
-            child: HexagonWidget.flat(
-              height: h,
-              color: Colors.orangeAccent,
-              child: Text('height: $h'),
-            ),
-          ),
-          Padding(
-            padding: EdgeInsets.all(8.0),
+            padding: EdgeInsets.all(padding),
             child: HexagonWidget.flat(
               width: w,
               color: Colors.limeAccent,
               elevation: 0,
-              child: Text('width: $w\nelevation: 0'),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildPointyExample(Size size) {
-    var w = size.width > 200.0 ? 200.0 : size.width / 2.0;
-    var h = 200.0;
-    return SingleChildScrollView(
-      child: Column(
-        children: [
-          Padding(
-            padding: EdgeInsets.all(8.0),
-            child: HexagonWidget.pointy(
-              width: w,
-              child: AspectRatio(
-                aspectRatio: HexagonType.POINTY.ratio,
-                child: Image.asset(
-                  'assets/tram.jpg',
-                  fit: BoxFit.fitWidth,
-                ),
-              ),
+              child: Text('flat\nwidth: ${w.toStringAsFixed(2)}\nelevation: 0'),
             ),
           ),
           Padding(
-            padding: EdgeInsets.all(8.0),
+            padding: EdgeInsets.all(padding),
             child: HexagonWidget.pointy(
               width: w,
               color: Colors.lightBlue,
-              child: Text('width: $w'),
-            ),
-          ),
-          Padding(
-            padding: EdgeInsets.all(8.0),
-            child: HexagonWidget.pointy(
-              height: h,
-              color: Colors.green,
-              child: Text('height: $h'),
-            ),
-          ),
-          Container(
-            color: Colors.red.withAlpha(100),
-            padding: EdgeInsets.all(8.0),
-            child: HexagonWidget.pointy(
-              height: h,
-              color: Colors.lightGreen,
-              inBounds: false,
-              child: Text(
-                'height: $h\ninBounds: false',
-                textAlign: TextAlign.center,
-              ),
+              child: Text('pointy\nwidth: ${w.toStringAsFixed(2)}'),
             ),
           ),
         ],
       ),
-    );
-  }
-
-  Widget _buildMore() {
-    return Center(child: Text('More examples will come with grids'));
-  }
-
-  Widget buildBody() {
-    var h = 300.0;
-    var w = 300.0;
-
-    var pad = 20.0;
-
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            Container(
-              margin: EdgeInsets.all(pad),
-              child: HexagonWidget.pointy(
-                width: w,
-                // height: h,
-                elevation: 4,
-                color: Colors.blueGrey,
-                child: Image.asset('assets/tram.jpg'),
-              ),
-            ),
-            Container(
-              child: HexagonWidget.pointy(
-                // width: w,
-                height: h,
-                color: Colors.orangeAccent,
-                elevation: 0,
-                child: Container(
-                  width: 1500,
-                ),
-              ),
-            ),
-          ],
-        ),
-        Row(
-          children: [
-            Container(
-              color: Colors.green,
-              child: HexagonWidget.flat(
-                width: w,
-                // height: h,
-                color: Colors.red,
-                inBounds: false,
-                elevation: 8,
-                child: Container(
-                  height: 50,
-                  width: 50,
-                  color: Colors.black,
-                ),
-              ),
-            ),
-            Container(
-              color: Colors.blue,
-              margin: EdgeInsets.all(pad),
-              child: HexagonWidget.flat(
-                width: w,
-                // height: h,
-                elevation: 4, inBounds: false,
-                child: Image.asset('assets/bee.jpg'),
-              ),
-            ),
-          ],
-        ),
-      ],
     );
   }
 }